+++ .github/workflows/python-lint.yml
@@ -27,6 +27,7 @@ jobs:
             ${{ runner.os }}-
       - name: Install dependencies
         run: |
+          cd ml_project
           python -m pip install --upgrade pip
           if [ -f requirements.txt ];
             then pip install -r requirements.txt;
