+++ .gitignore
@@ -103,7 +103,7 @@ celerybeat.pid
 
 # Environments
 .env
-.venv
+ml_project/.venv
 env/
 venv/
 ENV/
