--- metrics/metrics.json
@@ -1,1 +1,0 @@
-{"roc-auc": 0.9524, "recall": 0.7619, "precision": 0.9412, "f1": 0.8421}
