+++ ml_project/metrics/metrics.json
@@ -1,0 +1,1 @@
+{"roc-auc": 0.9293, "recall": 0.8636, "precision": 0.95, "f1": 0.9048}
