+++ tests/features/test_build_features.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from sklearn.compose import ColumnTransformer
 
-from src.features.build_features import build_transformer, extract_target, make_features
+from features import build_transformer, extract_target, make_features
 from tests.utils import create_fake_dataset
